+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 02/05/2022 - v3.4.1
+
+- (Add) Suggestion - Wait time before cure: Allow to set the number of layers to smooth transition from bottom to normal wait time (Defaults to 8)
+- (Fix) Tool - PCB Exposure: Pixels per millimeter was been set to fixed value instead of use printer lcd pitch, causing wrong dimentions on different from 50µm pitch
+- (Fix) Tool - PCB Exposure: Unable to run the tool when the display size information isn't available
+
 ## 02/05/2022 - v3.4.0
 
 - **Tools:**